disable: zram (use zramSwap), add: fastfetch Also removed vscode because im not using it.
Theres a big problem with AGS not picking up a font. I did some tests
with some nerd fonts and nothing worked. This may mean we need to
re-code the symbols, but it probably also likely means that AGS
can't pick up the fonts we need, due to whatever reason.