feat: vscode + a few bugfixes vscode is now entirely managed via declaration, though im still debating if this is truly the way I want to go moving forwards. After all, I still have a windows machine with vscode, and this would simply fragment my configuration (no im not getting Nix on MSYS2).
I patched a bug in the zsh that tried to persist the history file. It
would try and make a .local/state/zsh/zsh_history depending on the
directory zsh was opened in. That might be fine for most cases when
opening the terminal, but in vscode for example, itd create a .local/...
path in the working directory. Not quite what I want.

Firefox was being weirdly buggy and inconsisent. Maybe it was Plasma6's
fault by not having full protocol impl for browser://, regardless i
switched to ms-edge.