feat: finalize the arbitrary files module It's not perfect, but its come a very long way.
The idea is that you can do either:
```nix
my.system.file."/path/to/file" = {
  ...
};

my.users.USERNAME.file."~/path/to/file" = {
  ...
};
```

For the users, its intended to use a `~` in the path. It will be
correctly resolved to the right path and is only there for the sake of
looks. For system, the path must be absolute.

There aren't any checks to prevent bad values, and I won't be adding
them since this is not intended to be a bullet proof module. Maybe it
can be one day, but not today.