.github/workflow: add CI workflow to build configs
This will build all my configurations and run `nix flake check`. I'm hoping to eventually move the "build" logic into something in the `checks` flake ouput, which would allow `nix flake check` to handle it for me. #21 tracks that progress at the moment.

I had to add a disk-space free action in order to create enough space on the action runner. This could pose a problem for the future..