users/frontear: migrate off light to acpilight
This software has been abandoned and there are talks on Nixpkgs to remove it. I am also in support of the removal as it's going to slowly become outdated and it's better to pull the plug now than later.

We are using `acpilight` now, which contrary to the binary name `xbacklight`, does not _only_ work on X11. This is pretty much a drop-in replacement no special things to worry about here.