modules/nixos: change some nix.conf settings
Reduce `connect-timeout`, `download-attempts`, and disable `preallocate-contents`. At first these may all seem like bad choices, but they have their uses. Currently, `connect-timeout` can cause Nix to hang when it fails to connect since it waits for the timeout to be reached. This isn't for just one or two things, this is for _all_ network IO that occurs from Nix, which can block on a lot of unnecessary things, such as trying to update the local narinfo db when it doesn't do anything for our invocation. `download-attempts` follows in tandem to this issue, by causing further blocks on unnecessary tasks, and as a result we want to be slightly less forgiving under low network situations.

Disabling `preallocate-contents` is actually very important because it breaks BTRFS COW functionality and compression. Furthermore, the feature itself isn't actually that important on a modern system, so doing it ahead-of-time isn't bringing any major benefits. The negatives here outweigh the non-existent positives, so it has been removed.