hosts/laptop: remove bad recommendations for mounts
Preventing copy-on-write was a very bad idea. Originally my logic was that because I do not strictly care about the store surviving, since I can fully regenerate it through my configuration; it does not matter. Furthermore, COW is useless in the store since files are only ever created or deleted, it actually performs no tangible actions since there is never a "write" that happens on an existing path for a "copy" to occur. Unfortunately, I did not realize that disabling COW also disables compression, that means that my system has absolutely no compression on the store despite using `zstd:15`.. Next time I suppose.

Preventing compression on persist is up for debate though. I've removed it now because I don't have any strong evidence for or against it, especially since I haven't experienced the IO of compression from this installation given my mishap mentioned above. It may be worth doing but it may be too IO intensive, I'd need to perform real-world testing, which I haven't done. As a result, this recommendation may be back but it might not, it depends on what I discover in the future.