root|modules: update flake.lock, fix systemd messing with machine-id early
In the newest release of systemd (or the _newer_ release compared to the previous nixpkgs revision), a new `systemd-machine-id-commit.service` was introduced to help ensure `/etc/machine-id` was correctly saved from a transient fs like `tmpfs` to a real persistent device. This is for fairly obvious reasons, systemd relies on the machine-id to store logs and other system stateful information, and losing it each time defuncts those data.

Unfortunately, whilst well intentioned this service does nothing for my use-case because we already handle this ourselves (ignoring the `nixos-install` issue which has yet to be fixed). As a result, we can simply disable this service since there is enough assurance it won't be needed. Checking `systemctl --failed` shows no other services failing due to this change, which gives me the confidence in knowing this service is non-essential to anything else, and can be safely disabled.