.github/workflows: Improve pipeline with some parallel steps
In order to speed up the CI pipeline, we are parallelizing the builds. This does unfortunately re-perform a lot of the tasks, specifically the disk space free, which is arguably the second most lengthy step after the build itself. I cannot find out how to avoid this so at the moment this will be the best change we can do. If there are ways to "snapshot" an action's state and allow parallelization from that snapshot forwards, that could be cool. I know a VM/Container could do this, but I've no idea how to do that using GitHub Actions yaml logic.