hosts/laptop: use the xanmod kernel
Originally I used to use the latest standard kernel, however after 6.11 it would consistently fail large Nix evals. I could not determine the root cause but I did notice that my zram powered swap wouldn't fill up entirely, and yet oom-killer would fire.

So why the switch? I'm honestly interested in the performance patches, and I made sure to check that my system can still evaluate. Thankfully, it does! Regenerating a `nix-index` database used to fail with the latest kernel, but with the latest xanmod it does not. This is a good sign and as a result im merging in.