treewide: revamp the persistence module
Massive changes have been accomplished here. Gone is the large unwieldy bash script and in is a smaller and significantly simpler script that executes within systemd. This has one major benefit of parallelizing the mounting, which can have some pretty decent speedups. Another major benefit is that we have more fine-tuned control over the execution of our logic. This allowed us to insert it earlier in the initrd than even when NixOS activation occurs.

Alongside it comes a persist copy script, which will attempt to perform a copy of any file contents that exist during shutdown for the first time in order to create a trivial directory tree of the persisted entries. The main reason for this was simply to have a set of permissions to reference during mounting, but performing a mostly safe copy was good too. Regrettably there is no way to assure our copied content is complete, but that's simply not something we can handle at this stage anyways.

Permissions have been completely thrown out from the module. Instead, it is encouraged that permissions are managed on the filesystem itself. Is this bad for reproducibility? You could say, but I argue that permissions are stateful and shouldn't have been mingled here anyways. Software should be responsible for setting the permissions it needs and those permissions will be copied and respected through this setup. This is more than enough.

Since a lot of stuff within this module only works in a sane manner thanks to our persist being on "/nix", we're throwing out the ability to change it entirely. Most of our stuff works because "/sysroot/nix" is mounted in the sysroot early on alongside "/sysroot". Other mounts are not so lucky. Thinking of an implementation to force this to happen as early as possible whilst also thinking about the abstractions needed was simply too annoying, so I will not be considering it at all.