.github|nix: introduce nix flake check for CI
This will attempt to link together all of the relevant dependencies and force a build on them when `nix flake check` is invoked. This is useful to ensure our outputs are still building as expected, and sets the base for further checks down the line, such as run-time (with nixosModules being the first instance of this whenever we get to it).