users/frontear: remove editorconfig manual manipulation from neovim
This was incompatible and I misunderstood it's purpose. Instead, I've configured the vim specific knobs that corresponded to said settings. It's all kind of messy but it should work as expected.

On a different note, it may be a good idea to cleanup my neovim stuff. I'm just not nearly as concerned about it as I am with my NixOS configuration, but I really should be.