treewide: remove all uses of lib.pipe
o7. It served us so well for so long, but thankfully now we have an actual pipe operator that does the same thing but with slightly more flexibility.

Alongside this treewide change is an improvement to the cleanliness of the cachix packages. Overall I'm not terribly happy with this approach _still_, especially since it doesn't really properly consider what derivations are actually needed and which aren't, instead it just blindly grabs what _might_ be useful. This isn't the best approach and I'd like to deal with it properly instead. One idea is to maybe have a "toplevel" attribute in the modules that will contain all relevant packages. The flake attributes for cachix can then leverage this module option to expose the useful things only.