flake.lock: bump revisions + fix build issues
Seems that while I was gone Nixpkgs bumped to GCC 14. This has seemingly broken quite a few of the older archive utilities. Patool also broke, though I don't think it's related to GCC 14. The simplest fix was to rollback the stdenv for the broken archive utilities, and to pin a newer version of patool that was merged onto master but isn't yet in unstable. Once it rolls in I will remove the pin.