modules/nixos: replace Lix with Nix bleeding-edge, update some settings (and add some comments)
Lix is falling behind the normal releases of Nix, and I don't have the sufficient processing power to build it on my system. Furthermore, I would rather support the development of Nix directly, rather than loosely following behind Lix. Lastly, I do not appreciate some of the Lix community members and their attitude towards the Nix project and ecosystem. We all have our differences and while I'm also very disappointed with the state of affairs in the community, I do not think that an extremely hostile and rude attitude is the correct way ahead, and especially by diving the community into two subgroups. Nonetheless, that's an issue to rant about elsewhere.