modules/nixos: small nit fixes
Moved the `/var/tmp` configuration for the Nix daemon from the `/tmp` module to the nix module. Removed the desire for a zram writeback device since it would have no benefit on a tmpfs system anyways, unless I dedicated a drive to it. Lastly, changed the systemd-boot module to use `rec` in an isolated block.