feat: some minor preparatory cleanup for the persist module
Firstly, the persist.toplevel attribute exposes a list of attrsets containing a src and dst attribute, where the former points to the path in the persistent volume, and the latter points to the path it must materialize at on the root filesystem. This removes the unnecessary path construction burden that was tossed around all over the config expression.

A more minor change involves updating the lib.types.userPath to accept any string that matches the pattern `~/[^~]+`, or descriptively, a string that starts with `~/` and contains ne or more characters where none of them are `~`. This in particular prevents a possibility of providing a path like `~/~hello`, which would wrongfully be processed as `/home/name/home/name/hello`. Such a path would also naturally cause a matching issue in the configuration, which is a tad helpful for the rare cases where I type the wrong value here (which has yet to happen but it doesn't hurt to add it).