feat: update `flake.lock` and self-host the microsoft-edge derivation
It seems that Nixpkgs removed the Microsoft Edge derivation due to a lack of maintenance which could further lead to possible security vulnerabilities down the line. I'm not particularly sure what they are referring to with the second point given that a) I do not see any active CVE referenced anywhere in the discussion, and b) a concern that the bot was the only primary contributor/updater. I do not believe these reasons are sufficient, since a browser app like this really doesn't need much maintenance beyond updating intermittently.. We will be returning it to this repository until further notice, perhaps I may even simply commit myself to being the maintainer.