feat: update flake.lock, add version to wrapped nix derivation, fix libirecovery
The new `nixos-rebuild-ng` needed `nix.version`, and our wrapper did not expose it by default, so I've gone ahead and propagated the version from the original nix derivation upwards.

`libirecovery` has an issue in creating it's udev rule that caused build failures. A PR has been proposed that indeed fixes the issues and causes it to build fine, and so I've opted to overlay it into my system. This will be removed when the PR has been merged into Nixpkgs.