feat: bump flake.lock, remove temporary home-manager fix, fix zsh module
Home-manager fixed the issue with the gcroot disappearing, and `sd-switch` now functions as expected.

The home-manager zsh module originally used to expect a relative path, which forced us to do weird shenanigans to fix it. It has since been changed to take an absolute path, so we could drop our weirdness.