feat: introduce a GNOME specialisation and cleanup modules
GNOME is intended to be my work-only environment, where I focus solely on school. It's great because of it's calendar and integration with my online accounts OOTB. It's also really pretty and simple, which is great because I don't want to waste time trying to make it work right, and Sway just isn't usable enough for me.

The microsoft-edge module got revamped into being a generic chromium module, as I've chosen to use `google-chrome` for the GNOME session and not `microsoft-edge`. This is mainly to prevent issues with different secrets providers causing a mess in the browser's settings, and to help further separate work and personal stuff.

The libreoffice module has been cleaned up to not have these random extra settings, as per #31

Some applications that came set up pre-installed for my user have been moved into desktop specific configuration. This is for the same reason as above, separating personal and work stuff.