feat: separate sway's fs derivations to isolate rebuild propagations
Fixes #41. I honestly wish I didn't have to do it this way, since now I have to repeat unnecessary boilerplate in each `fs/*/default.nix`, but this seems to be my only option.

I did actually try `__contentAddressed`, but given that it's an experimental feature, and with no actual guarantees that it would work in the same way, I figure it's not worth diving into. If it stopped working right, I'd have to rip the derivation apart anyways, so doing it earlier seems better.