feat: new user service, cleanup existing services, fix sway exit
Both `waybar.service` and `foot-server.{service,socket}` were a tad extra with options that existed in the upstream unit. They aren't bad, but in my unique setup I don't really need them, so I've stripped them down to a minimal unit. I'm also explicitly stating which slice the services should run in, to help divide up resource usage in the live system. I chose not to use the `uwsm`-specific slices like `app-graphical.slice`, `background-graphical.slice`, as I don't particularly see why they exist in the first place, and why I should bother with them.

I moved `swayidle` to it's own module and set it up with a user service, following the same principles as earlier. I also moved the configuration of `swayidle` to it's own file, which is passed in by cli args to the program. Neat!

Finally, I changed the sway exit behaviour to invoke `uwsm stop` rather than killing the service directly. This was a documented request in the `uwsm` man pages, so I figure why not. Admittedly I almost never actually this keybind anyways, but it's still a good change to make.