feat: massively simplify sway configuration and cleanup some files
My waybar configuration has been simplified extremely, simply because I honestly didn't need a lot of the stuff I was doing on it. I've also tried to re-organize the modules in the order that they are on GNOME, purely for aesthetic reasons.

I've also changed a setting in waybar that was adding extra margin to the bottom of the bar, making it stand out since my sway also had `gap_out 4` set. There is now no margin on hte bottom, which makes it look more evenly spaced.