feat: set fonts, cursor, and improve icon/theme configs
We now set default fonts through `fonts.fontconfig.defaultFonts` so that Qt apps and other apps which aren't explicitly configured can use the fonts. We have a new emoji font to go with this.

The cursor has been set to `bibata-cursors`, particularly the modern classic variant, as it looks really nice. This automatically adds the relevant configuration to GTK as well. Sway was manually changed to have this too.

Qt configuration has dropped `QT_QPA_PLATFORMTHEME` as it's not necessary with the Adwaita style that I'm also using. The [Arch Linux wiki](https://wiki.archlinux.org/title/Uniform_look_for_Qt_and_GTK_applications) was very helpful for learning more about this.

Both `dunst` and `rofi` have the icon theme and font set respectively. In particular with `rofi`, the default font size has been shrunk a bit, just so that it takes less space on the screen.