feat: partial cleanup of `modules/nixos/system`, create `modules/nixos/services`
The difference between a "service" and a "program" is that a service exists entirely in the background, and acts independently of the user to provide a feature/functionality to the system. A program is an interactive on-demand feature/functionality on the system. A program _may have_ a background service, but it's existence is intertwined with the software.

Note that this new belief will require a partial refactor of the modules. `dunst` might need to become `my.services.dunst` rather than `my.programs.dunst`. That's a future problem though.