feat: cleanup `direnv`, `git`, and `gnupg` modules
Parts of these changes are part of the ideas from #43, focusing on home-manager interop. The only options that are exposed are the critical ones that my user uses. Notably, `zsh` was exempt from the cleanup (and `neovim` but we don't talk about that right now). The main reason for that is because those require more testing and iteration for now.