feat: the great unwrapping ft. dunst
Part of #43. I also went ahead and disabled the `icon_path` shenanigans that home-manager was doing, because I wanted to make use of the newer standard that dunst was trying to go for. Supposedly it will become the default too, so even better.

I will admit the color difference is jarring, I was used to the blue for low, orange for normal, and red for critical. I may manually tweak those colors back possibly, but for now I won't.