feat: set `stylix.autoEnable = false` for home-manager
Default enabling this enabled every target (except for a select few according to their documentation), and some of them introduced an overlay in `nixpkgs.overlays`. This caused an assertion warning during evaluation due to the fact that I have set `home-manager.useGlobalPkgs`, which inherits the `pkgs` from NixOS, and therefore only uses the NixOS `nixpkgs.*` module settings, not it's own. I've just gone ahead and disabled it, I'll worry about it when I need those targets.