feat: use icons provided by the gtk theme for bottom bar
This is soo hacky and seriously ugly, but it works. The file names are based off of the `Icon` attribute from the `.desktop` file. I tried to see if I could access the property from the derivation, but I could not, and I honestly didn't look further. It's honestly really disappointing that waybar can't just request a particular gtk icon. I wonder if there's anything I can do to extend waybar, or maybe it's time to replace it? It also might be worth giving this some kind of function..