feat: remove subflakes `flake.lock` and add to .gitignore
These flakes are linked directly to this main flake through an `inputs.self` (probably an awful naming choice considering `inputs.self` is an actual attribute that a flake can use..). This means that they do not need their own lockfiles, and whilst one will generate, it's completely worthless to keep track of.