feat: improve reliability of persist module, add remount service
The services are now all `Type=oneshot` to ensure they complete before the target they are `RequiredBy`, which helps guarantee more consistency in their execution step. I did a quick benchmark with 5 runs pre-changes, and 5 runs post-changes, and found that speeds are pretty much identical. This tells me that the changes aren't adding any noticable delay to startup, which is extremely good.

I've also introduced a new remount service that attempts to add the `x-gvfs-hide` mount option to all the bind mounts. This happens early enough that it doesn't conflict with anything, and fixes the long-standing issue at #40.