feat: deprecate cachix entirely
I have not made use of this cache in many months now. I just don't get much use out of it right now, and I don't see a future where I will be getting that much use out of it either. If I do end up needing it, I'll re-introduce _some_ of these things back, but not most of them (the module, for example, is not coming back as it is right now).