feat: bump flake.lock and resolve minor issues
It seems that home-manager is deprecating their automatic insertion of the ssh config.. defaults. That's right, every single option they set was already a default of the program, yet was declared because.. reasons. The temporary drop-in will stay until they remove that option, after which it will be removed.

My upstream PR for systemd-oomd was merged, so I can drop my temporary drop-in for that as well.

Finally, nix 2.24 is gone from nixpkgs, so I've bumped the nix binaries in `nix-benchmark` to include nix 2.31. I've also uncommented the git revision of lix, I don't know why I didn't have it enabled prior.