feat: create mount units in the initrd to orient persist mounts
This is a partial change aimed at making the implementation of #33 easier. Alongside this, I've taken the liberty to remove more of the fluffy text in the service comments, and leave more meaningful comments describing what it does and why certain choices were made. Some cleanup was done to unit conditions, one notable one was dropping the requirement of `!${path.dst}` from `ConditionPathExists=` in the initrd units, since once #33 is implemented, destination paths might partially exist and could cause issues with service firing. Another change was fixing the usage of `cfg.volume` in a lot of places, since #33 would introduce new locations that may be suffixed with specialisation names.