feat: devshells for all in-tree packages without subflakes
It was getting a little too hacky and messy trying to work a subflake that depended on a higher flake, whilst also making sure to not leak the useless `flake.lock` that it would create. Instead, I've opted to manage devshells at a higher level for all the in-tree packages. This is a significantly better way to do this, and I'm a lot happier with this approach.

Coming alongside this is a cleanup to where `package.nix` is located for these packages. They will all be at `pkgs/<name>/nix/package.nix`, so resolution with `lib.flakes.mkPackages` is less convoluted. There is also some additional cleanup of the derivations themselves to make them more consistent. This cleanup has resulted in a hash change, but otherwise the contents are identical. Hopefully this is the last time I do anything like this.

Finally, I've killed the `nix/devShells/bin` directory entirely. I left it initially because I considered wanting the `install` command, but I don't really need it.