feat: init `persist-make` as a replacement for `persist-helper`
I noticed that after setting up the initrd mount units, `persist-helper` was only really doing the work of `createDirectories`. Furthermore, as noticed in #50, it was quite slow. In order to speed it up, I have written a Rust application that _only_ does the job of `createDirectories` and nothing more. I've also taken the liberty to completely get rid of the logging, to speed up timings even more.

Benchmarking reveals a massive performance improvement. Regardless of path size, the Rust application takes 1.3ms~, whilst the Bash script can take 20ms~ on a shorter path, to 100ms~ on a longer one. The functionality is pretty much identical and can be a drop-in replacement, with some minor tweaks to the arguments.

Note that this Rust application has no error checking. This is a deliberate choice to keep the code simple, since our units have a plethora of condition checks to protect against problematic execution.