feat: make user persists unique by default
The microsoft-edge paths being shared was kind of broken, and it made me realize that it's a lot better to simply default to uniqueness for user paths, and exclude certain applications, rather than the other way around. Eventually I should comb over all my modules and set the paths explicitly, but that's a future issue.