feat: revamp the vscode module, use fhs for vscode
Yes, purity is dead..

Just kidding, but in all honesty this is a similar approach to what I've already been doing with my Neovim configuration. It's not 'pure' by any means, but it's also not intended for anything other than making the extensions work their magic. Actual compilation has, and always will be powered by a devshell. That isn't going to change.