feat: fix root being world-writable
I honestly cannot believe that I never noticed this until now. This whole time my `/` was world-writable. I guess a big part of the reason why I didn't notice it was because every single directory or file that was created on the root was usually `0755` or `0644`. This prevented the `1777` permissions of `/` from leaking into anything else.