feat: add `zed-editor`
I'm actually very against this editor atm. I'm only pushing this commit because it's almost 5 am, and I want to sleep.

One major issue with this application is that it attempts to download language servers via `node`. Yes.. `node`. It apparently also provides no way to just use language servers from `$PATH` either, making it impossible to work with 99% of languages whose extensions aren't configured to pull the LSPs from `npm`. The _only_ fix is to wrap it in an FHS environment and lock up the nasty stuff in there. This is frankly really disgusting and I don't think I want to use an editor that works like this.

I do have to say though, Zed looks a hell of a lot nicer than VSCode. It's significantly more minimal, and that's sort of made me fall in love with it. Alas, this FHS stuff is nauseating, and unless there's a nicer fix, I'm tempted to just completely erase it from my repo. I'll leave it alone for now, maybe the features are in the works.