feat: drop Sway and switch to Niri
This had been coming for a while, but it's only today that I finally got the chance to sit down and make the change. It's actually quite nice that majority of my configuration was transferable with no major issues.

Niri fulfills the scrolling + tilling combination so much better than what Sway w/ Papersway could. This is definitely going to be my final compositor experiment, I don't see myself switching to anything else after this.

Most of the stuff in this PR was simple transitions. There wasn't much attempt to tap into Niri's flexibility, that's for the future.