feat: drop Zed
I'm not going to do into a whole rant about the problems with this editor, instead I'm going to focus on the problems with using it on NixOS.

Extensions do not install automatically, despite declaring them in the home-manager module. It only "encourages" Zed to install them after startup.

The `settings.json` and `keymaps.json` are both impurely placed on the system by home-manager. I assume there must be a reason though I did not dig into it.

Extensions use some downloaded `node` binary to download and install LSPs, kind of like `mason` for Neovim. This is a no-no for NixOS.

Using this editor is more trouble than it's worth. VSCode by comparison is a lot easier to work with, since tooling has been built around it's existence in the Nix ecosystem. Maybe Zed will have the same tools in the future. For now, it's not worth it.