feat: drop the `fs` derivation, write configs in Nix, new programs
The config is still just a big chunk of lines, but it allows me to interop values from the rest of the config into it. I've also taken the liberty to re-organize a few modules, as well as add 2 new programs to my little bar at the bottom, Legcord with Discord's icon, and OBS Studio.