feat: bring legcord to GNOME
Long story short, I use Discord to occasionally dump notes from class (usually off of my iPad) and it's a bit of a hassle to try and access the notes from my laptop when I need them. I considered trying to use Obsidian, a markdown note-taking application, but setting up a free sync that would just work on both my iPad and laptop proved to be a bit too much of a hassle to explore. As such, I'm just bringing Discord (via Legcord) to my GNOME specialisation. To try and keep myself from being needlessly distracted, I'm using a different account that can still access the notes, but does not have many of the other servers.

I might explore using Obsidian in the future, but for now it's a bit too involved for me to waste my time on.