feat: the beginnings of a more organized modules directory
See #53 for the details surrounding this change. This is the initial commit for this, which means that there is a LOT of work that needs to be done. Rather than trying to do it all in one commit I wanted to split it up across multiple (which is the correct way to be doing this stuff anyways).

For starters, `lib.mkModules` need a massive overhaul so that it can perform most of the heavy lifting for module files resolution. This will allow me to drop all of the `module.nix` files that exist only to import `options.nix` and `config.nix`, simplying my structure considerably.

After the overhaul, I need to move modules around to have them match the expected structure. This is definitely going to be the easiest part.