feat: overhaul `lib.mkModules` to dynamically resolve modules
A full implementation of #53. I'm not going to explain the code itself because I feel that it is sufficiently self-documenting, but I will elaborate on some smaller details.

`lib.mkModules` is intended to enter from the top-level `modules` directory rather than for specific modules. This is because it takes full control of resolution by making the assumptions laid out in #53. The `nixosModules` exposed by my flake have been updated to reflect this.

I've also changed a lot of path construction in `lib.flakes` from using string interpolation to using concatenation. The main reason for this is because concatenation will respect the original value type and possibly avoid the idiotic flake copy semantics. To explain this further, when you try to interpolate a path into a string via `"${pathValue}"`, Nix will copy the path into the store. This can be a very expensive operation depending on whether it's a big file or not, and it's really annoying that it does this. On the other hand, concatenation will respect the original path type and concatenate it into a new path value, without performing the stupid flake copy behaviour. This avoids unnecessary I/O operations and potentially fragmenting files by grabbing them in their isolated contexts and not as part of the flake that they evaluate in.