feat: improve the nixos-clean helper script
Motivated by #49. This makes use of `zparseopts` which requires that the interpreter be changed from `bash` to `zsh`. Small other improvements were made, though the script functionality remains identical to before.

I'm considering creating a Rust multi-call binary instead to implement these applications, as otherwise their codebase is going to duplicate a lot of stuff (such as the `dry-run`, `verbose`, and `help` options, the log levels, etc). It might be better off to do that.