fix: add `TODO` note for recursive module importing
I'm honestly a little surprised I didn't catch this originally, especially since it was a core part of the specification I laid out in #53. To my credit though, I have faithfully followed that specification, and this wouldn't be the only snippet that doesn't have error-checking, so I'm not entirely sure if I should bother with this at all.