fix: update flake.lock
Nothing noteworthy to migrate this time around. I suspect most of the big changes have already made it into the release-cycle, and the final few are still churning in staging. Release cut-off is November 30th, only 20 days (3 weeks) away.

I would've loved to contribute to the release cycle, but alas, school is a full-time job in and of itself.