feat!: use DankMaterialShell for Niri
This was a rather big change, and was also rather spontaneous. Overall everything should be set up perfectly fine, but there are a couple things to note:

1. The initial settings from `default-settings.json` do not apply. The
   program needs to be restarted for them to work. I suspect this has
   something to do with DMS copying but not changing the settings. This
   will need to be fixed.
2. Night mode does not seem to toggle on if auto mode is enabled. This
   is peculiar but it's not super high priority so I don't really care
   too much.
3. DMS has it's own Polkit agent, I could drop the GNOME one for this.

Overall it looks nice, and it implements many of the things I didn't bother with. Once upon a time I would've wanted to write this stuff myself, but not these days. I'd rather just have something that works.