fix: sync DMS setting pre-start so that changes are visible
Before this fix, DMS was simply copying the default settings and then not parsing them. As such, the configuration of DMS was stuck as the default, and a restart of the service was necessary to see the changes. This would become a problem on next boot, since I use a tmpfs root, so the solution was to perform the copy ourselves rather than letting DMS do it.

On a separate note, I should investigate what would happen if the `settings.json` file was managed directly, rather than through this indirect approach. At the moment it doesn't really matter, since tmpfs root + destructive copy means that imperative changes won't stick around for very long, but it's worth investigating to see if I can avoid this indirection. Of course, the one thing that would suck about this would be losing the ability to make an imperative change, then checking what JSON key it corresponds to, and setting it in my Nix configuration. Maybe I should leave it as is, just for this fact alone.