fix: repair an irreversible destruction of git history by a flake input
Just awful. It's honestly kind of annoyed me that the author did this. They've apparently just wiped the entire flake clean, and also wiped all of the git history pertaining to it. I cannot understand why they'd do that, but I have to say that's not only ridiculously irresponsible, but extremely annoying.

Well anyways, I'm fortunate enough to have had a copy of it in my nix store, so I've gone ahead and grabbed it from there and referenced it in-tree. It will be removed very shortly, but I needed to make this fix so that I could push the staging cycle through.