feat: significantly reduce the amount of runs done in `nix-benchmark`
I feel like 20 runs is mildly excessive. It takes up a lot of CI time and from local testing, doesn't seem to really improve the accuracy of the final time measurement. At the very least, the accuracy is negligible compared to the runtime cost, so I'm dropping it by many factors. Warmup runs are going from 5 -> 4, and full runs are going from 20 -> 8.